Created the OfficeHandler class and used it in index.ts
